+++ src/demo/utils/demos/scroll-up-only.vue
@@ -1,0 +1,8 @@
+<template>
+  <TvScrollTop :show-on-scroll-up="true" />
+</template>
+
+<script setup>
+import { TvScrollTop } from "@todovue/tv-scroll-top";
+import "@todovue/tv-scroll-top/style.css";
+</script>